Ignore keyword arguments in SSA logic
Assumes that when we visit an Arg node (as part of a Call), this should
always be in the "reading" context (so it will just replace the
reference with the latest SSA value, but not cause any changes to the
state of the SSA mappings).  Then, we can just discard the changes and
reuse the original keyword (since these are not variable names that
should be affected by SSA).